Modified textUI class to make it follow single responsibility principle
